a bit changed files

+++ src/main/java/model/entity/Entity.java
@@ -4,21 +4,21 @@ import java.io.Serializable;
 
 public abstract class Entity implements Serializable {
     private static final long serialVersionUID = 1L;
-    private int id;
+    private long id;
 
     public Entity() {
 
     }
 
-    public Entity(int id) {
+    public Entity(long id) {
         this.id = id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
